all required changes applied

+++ src/controllers/userController.js
@@ -217,11 +217,11 @@ export const deleteUser = asyncErrorHandler(async(req,res,next)=>{
         return next(err);
     }
 
-    console.log(userExist);
     
+     // delete photo from the cloudinary
 
-    // delete photo from the cloudinary
-    await deleteImageHelper(userExist?.photo?.public_id);
+    if(userExist?.photo)  await deleteImageHelper(userExist?.photo?.public_id);
+    
     // deleting user
      await User.findByIdAndDelete(id);
